(split)bug: 25228, 25329
When an account is renamed, update zimbra.mailbox table to set "comment" column to the new email address and set "last_backup_at" column to null.  This causes the next backup (standard or auto-grouped mode backup) to force a backup of the renamed account.

Restore using either old or new email uses the latest backup for the new email address if no -lb option (backup using specific label) is used.  The restored account will have the new email address.  To restore the data prior to the rename, you must run "zmbackupquery -a <old email>" to determine the last full backup under the old name and restore using -lb option.  This will still end up restoring the account to the new email address because the rename operation is recorded in redolog and is replayed.  To end up with old email address, you have to stop the redolog replay before the rename operation, using the various -restoreTo* options.  (But note determining the exact point the account was renamed is not easy.  You have to run RedoLogVerify on the redologs and find the timestamp of the RenameMailbox operation.)

http://bugzilla.zimbra.com/show_bug.cgi?id=25228
http://bugzilla.zimbra.com/show_bug.cgi?id=25329

Copied from Perforce
 Change: 87001