bug: 21244
When handling deferred index updates, the Mailbox will attempt to detect situations where the index itself is corrupt and backoff index-retries, instead of just repeatedly trying to index all the deferred messages.

VERY IMPORTANT RELEASE NOTE DATA:  The server will log a message that starts with "Possibly corrupt index" at the WARN level to the zimbra.index logger if it detects a corrupt index.  The admin is responsible for monitoring the logs for these situations and correcting as appropriate (reindexing the mailbox, fixing disk space issues on the index volume, etc).  

http://bugzilla.zimbra.com/show_bug.cgi?id=21244

Copied from Perforce
 Change: 87482