(split)bug: 25095
Add two new LC keys to control the loading of the standard message converters: zimbra_converter_enabled_uuencode and zimbra_converter_enabled_tnef.  They default to "true"; if set to "false", their respective converter is not loaded and UUENCODE/TNEF decoding is not performed on the fly.

http://bugzilla.zimbra.com/show_bug.cgi?id=25095

Copied from Perforce
 Change: 87755