(split)bug: 18855
The update invite email that is sent by the server was not setting SENT-BY field correctly in ORGANIZER property.  It was always retaining the existing SENT-BY value.  Instead it should be set to the authenticated user when changing organizer to another user, or cleared when authenticated user is changing organizer to self.

http://bugzilla.zimbra.com/show_bug.cgi?id=18855

Copied from Perforce
 Change: 88496