(split)bug: 26006
During ics export stream the output stage.  We still have to query the mailbox and get the list of appointments loaded in memory, but we can at least avoid the further memory penalty of generating iCalendar version of the data as a byte array before writing it to the client socket.  Large description texts of VEVENTs are loaded during the streaming stage, so they won't add to the memory pressure.

http://bugzilla.zimbra.com/show_bug.cgi?id=26006

Copied from Perforce
 Change: 89245