(split)bug: 23985
On deeper inspection, the "should we force the index" logic actually turned out have been broken a while ago by some unrelated cleanups....fixed it so that it worked the way it was originally intended -- which was to NOT force the index if there was a tag clause (e.g. is:unread).  

http://bugzilla.zimbra.com/show_bug.cgi?id=23985

Copied from Perforce
 Change: 90428