(split)bug: 26318
Return mailbox size changes in JSON <notify> and <refresh> as a list of objects, e.g.

   mbx: [ {s: 324897}, {s: 89923, acct: "acct-id-..."} ]

Updated zcleint to handle this.  Note that in the case that the ZMailbox points at a mailbox other than the authenticated user's, mSize may not be updated correctly.  (Is there a way to get the authenticated userid and the requested userid in zclient?)

Slightly better synchronization defense in delegated notification processing.

http://bugzilla.zimbra.com/show_bug.cgi?id=26318

Copied from Perforce
 Change: 90447