bug: 24559
integrated from franklin.
there needed to be some more logging in ZimbraMsgStore::GetProps() to see what we are returing for PR_ENTRYID.  this is the last call that is made before the OOM exception is returned.  OOM probably uses the value returned by this property to determine MAPI_E_NOT_FOUND even though the property is there.
also added more logging to MSProvider::CompareStoreIDs() to dump the wrapped versions of the entry ids that we are returning for further comparison when this problem happens.

http://bugzilla.zimbra.com/show_bug.cgi?id=24559

Copied from Perforce
 Change: 90997