(split)bug: 26821
Do all database operations of a mailbox restore in a single db transaction to benefit from atomic commit/rollback.  The only exception is the CREATE DATABASE of mailbox group database if it doesn't already exist.  This is done at the beginning, so all INSERT/UPDATE/DELETE statements are still contained in a single transaction.

http://bugzilla.zimbra.com/show_bug.cgi?id=26821

Copied from Perforce
 Change: 92346