(split)bug: 25714
When removing a subscription, the RFC tells the client to send *only* a <presence type="unsubscribed"> request to the server instead of sending both an IQ.set and a presence unsub....(this is the opposite of the subscribe protocol, which tells the client to send both).  We were doing the wrong thing which was generating spurious (and incorrect!) roster updates to the client.

http://bugzilla.zimbra.com/show_bug.cgi?id=25714

Copied from Perforce
 Change: 92860