bug: 26422 - Filter trying to "File into folder" a shared folder fails
ZimbraMailAdapter now supports filing a message into a mountpoint.
This is done in a kind of roundabout way.  When the destination folder
is a mountpoint, the filter code does a REST POST to the mailbox instead
of calling Mailbox.addMessage().  I filed bug 27417 to address the
workaround.

http://bugzilla.zimbra.com/show_bug.cgi?id=26422

Copied from Perforce
 Change: 94491