bug: 16601
added keystore-backup and keystore-restore ant targets,  handy for backing/restoring keystore that are not generated by the localhost-ssl-cert target in dev env (reset-the-world/reset-all always regen the keystore).   Note, keystore-backup and keystore-restore have to be run manually only when as needed, they are not incorporated into other targets.

http://bugzilla.zimbra.com/show_bug.cgi?id=16601

Copied from Perforce
 Change: 94959