(split)bug: 27533 - Legal intercept does not work when message is streamed from disk
Change 94967 exposed a bug in ParsedMessage.fileMoved().  When a message
was in memory and fileMoved() was called, it would set mRawFile to a non-null
value.  As a result, other ParsedMessage code started treating it as a streamed
message.  Changed fileMoved() to only update mRawFile when it's non-null.

http://bugzilla.zimbra.com/show_bug.cgi?id=27533

Copied from Perforce
 Change: 95113