(split)bug: 27457 - BlobInputStream.skip() and reset() should use the read buffer
Updated BlobInputStream so that the skip() and reset() methods only
reposition the file pointer and don't do any I/O.  Rewrote the read code
and removed the assumption that the file descriptor is positioned
immediately after the read buffer.  Updated read(byte[]) so that it buffers
when the number of bytes to read is smaller than the buffer size.

The only remaining inefficiency is that we're doing a seek() before every
buffered read.  This is done to ensure that the file descriptor is in sync
with the logical position.  We can look into optimizing this code later,
if we determine that no-op seeks create enough unnecessary I/O to
warrant the added complexity.

http://bugzilla.zimbra.com/show_bug.cgi?id=27457

Copied from Perforce
 Change: 95205