(split)bug: 26568 - stats not being fully initialized at startup can cause NPE
For the life of me, I can't figure out what caused the NPE.
ZimbraPerf.POP_TRACKER is statically initialized, so it should
be there when someone accesses the ZimbraPerf class.
Changed all the public static vars in ZimbraPerf to be final,
to ensure that no one changes them behind the scenes.

http://bugzilla.zimbra.com/show_bug.cgi?id=26568

Copied from Perforce
 Change: 95249