bug: 27570 - cnvcache needs to be cleaned up -- dies when 32,000 reached
Removed wacky code that persisted the state of the conversion
cache to a file and read it into a data structure.  The old code
was error-prone because it assumed that the content of cache.dat
and /opt/zimbra/convertd/cnvcache would always be in sync.
If an unexpected shutdown ever occurred, we'd end up with files
in cnvcache that would never get cleaned up.

The timer task now iterates all the directories in
/opt/zimbra/convertd/cnvcache
and deletes the ones that are more than a week old.  This happens
once every two hours.

http://bugzilla.zimbra.com/show_bug.cgi?id=27570

Copied from Perforce
 Change: 97484