(split)bug: 22913
Check the new invite permission when receiving emails with calendar invite parts.  If recipient mailbox has not granted permission to anyone, process the calendar invite.  This mimics previous behavior when no permission check was done on anyone.

If permission was granted to some users and the sending user is not on the list, do not process the calendar part.  Email is still received but no appointment is created/updated/cancelled.

For invite emails sent on-behalf-of, use the address in Sender header for permission check, before looking at From header.

http://bugzilla.zimbra.com/show_bug.cgi?id=22913

Copied from Perforce
 Change: 98537