(split)bug: 21580 - Update SendMsg and AddMsg to stream from disk
Updated SaveDraft code to stream from disk.  Fixed a glitch in
MailItem.setContent() so that it returns the new blob instead
of the old one.

Updated ParsedMessage accessors to not read
the byte[] into memory unless the caller asks for it.  This
involved adding MimeMessageOutputThread as a gross
hack, since JavaMail doesn't provide InputStream access
to the full content of a MimeMessage.

Tested sending a 100MB attachment between two
accounts and confirmed that the heap size does not spike.

http://bugzilla.zimbra.com/show_bug.cgi?id=21580

Copied from Perforce
 Change: 99843