(split)bug: 25449
If we get an IOException trying to parse an HTML part for defanging, it probably means that the <META> Content-Type had an invalid charset.  ("unicode" is the regular offender.)  In that case, fall back to parsing the part with the user's default charset or the system charset.

http://bugzilla.zimbra.com/show_bug.cgi?id=25449

Copied from Perforce
 Change: 100338