bug: 28582
integrated from franklin.
make sure that we use the DDL wrapper for folder objects when copying messages and copying folders.
if we don't, then the entry IDs of the copied/moved folders/messages will change.  this is bad for the ID mapper and this is bad for change record generation.

http://bugzilla.zimbra.com/show_bug.cgi?id=28582

Copied from Perforce
 Change: 100588