(split)bug: 28666
Fix wildcards with to:/from:/cc: queries -- turns out they've never worked.

When parsing the query string for TO/FROM/CC queries, if the query string is a wildcard don't use the custom analyzer to generate the query phrase from the input string.

This might not be the right thing to do for people using custom analyzer plugins -- but it will work OK for the standard case.

http://bugzilla.zimbra.com/show_bug.cgi?id=28666

Copied from Perforce
 Change: 101037