bug: 28546
p4 integrat -b FRANKLIN_to_main @101530,101530
Rewrote how we do push.  Now we do accurate check of whether a mailbox change is relevant to client push.  If not we continue to wait (without resuming continuation).  To be iphone friendly we now persist previous Ping's heartbeat rate and folder list, so that when we receive an empty Ping we can use saved parameters.

http://bugzilla.zimbra.com/show_bug.cgi?id=28546

Copied from Perforce
 Change: 101531