(split)bug: 28663
When creating a new appointment during email delivery of calendar invite, ignore the alarm specified by the organizer.  Add a new one based on the preference of the attendee.

The preference is zimbraPrefCalendarApptReminderWarningTime attribute.  If it's unset, no reminder is added.  If it's set, reminder is set to that many minutes before start time if appointment is not an all-day appointment.  For all-day appointments, the reminder is always set to 18 hours before.

http://bugzilla.zimbra.com/show_bug.cgi?id=28663

Copied from Perforce
 Change: 102289