bug: 27240
We were constructing the URL for the classloader by catting strings together with a prefix of "file://".  This turns out to break on windows.  I'm not sure if it was always broken but at least it breaks with 1.5_15.  Fix is to first construct a File object with the conf/msgs directory path, and then get URL from File.toURL().  This way it's up to jdk to figure out the right format for each platform.

http://bugzilla.zimbra.com/show_bug.cgi?id=27240

Copied from Perforce
 Change: 104100