bug: 28348
Added three new localconfig keys to control which text/calendar parts are auto-added to calendar during email delivery.


* calendar_allow_invite_without_method (default = false)

Calendar clients should set "method" parameter in Content-Type header of text/calendar part.  If a text/calendar part doesn't have the parameter, it means the calendar part is only an attachment and should not be added to the calendar.  This is the default behavior.

When set to true, treat the calendar part without method parameter as calendar invite and add to the calendar.  (This is what we used to do before.)


* calendar_allow_invite_with_publish_method (default = false)

By default only invite with method=REQUEST are added to the calendar.  Some clients may send the invite with PUBLISH method instead.  Clients shouldn't do that, but in case such a client is encountered, set this config key to true to auto-add PUBLISHed invites.


* calendar_allow_forwarded_invite (default = true)

When an attendee forwards the invite email to another user, this config key controls what happens on the recipient side.  By default the invite is added to the calendar.  Set it to false and it won't be added.


Any text/calendar part that is not auto-added will show up as an attachment of the invite email.

http://bugzilla.zimbra.com/show_bug.cgi?id=28348

Copied from Perforce
 Change: 104472