bug: 29596
Added localconfig key calendar_ics_export_buffer_size.  This is the size of the in-memory buffer used during ics export.  If data being exported is smaller than this size, it is buffered in memory before being written out to the client over the network.  If data exceeds this size, it is buffered to a temporary file.  The buffering is done to avoid locking the mailbox during network write, which can be slow depending on the client or the network connection.

This key defaults to 131072 bytes (128KB).

http://bugzilla.zimbra.com/show_bug.cgi?id=29596

Copied from Perforce
 Change: 105370