bug: 29563
Avoid cross-database transactions in SQLite by writing to a MBOXGROUP<N>.MAILBOX table instead of a ZIMBRA.MAILBOX table when stats change on the mailbox.  When we load a mailbox, we make sure that the two rows match and update the ZIMBRA.MAILBOX row if they don't.  Every 10 changes, we batch up all the changed MBOXGROUP<N>.MAILBOX values and write them all to ZIMBRA.MAILBOX.  As a result, a few mailboxes' stats may be a little out of date for some admin commands, but the system will run off the correct values (since it fetches for the MBOXGROUP authoritative values) and the problem should self-correct.  This feature is turned on by the new "debug_defer_mailbox_updates" config key (defaults to FALSE, of course).

Perf gain is significant, bringing SQLite into line with derby though still significantly slower than MySQL.

http://bugzilla.zimbra.com/show_bug.cgi?id=29563

Copied from Perforce
 Change: 106587