(split)bug: 13821 - Mail Filters with POP Accounts not working
Mail downloaded from a POP3 data source is now affected by
filters.  Pop3Import is now no longer a subclass of AbstractMailItemImport,
because we now have another difference in behavior between online
and offline.  The offline client only applies mail filters when messages
are delivered to the INBOX folder.  In the online case, a data source
can be associated with any folder, so we want filters to always apply.

http://bugzilla.zimbra.com/show_bug.cgi?id=13821

Copied from Perforce
 Change: 106679