(split)bug: 29895
A call to MailItem.rename() that changed both name and folder was writing the new FOLDER_ID while doing the name-change part.  This caused the subsequent folder-change part to no-op since the folder had already been changed.  As a result, for move-renamed folders the PARENT_ID was never modified and the two columns diverged.

We address this issue in a number of ways:
  - folder cache instantiation is done based on FOLDER_ID, not PARENT_ID
  - DbMailItem.saveName() updates PARENT_ID for renamed folders
  - DbMailItem.move() now always updates the target row rather than no-op'ing when the FOLDER_ID is unchanged
  - when we load folders, we reset PARENT_ID to equal FOLDER_ID if old evidence of this bug is detected

http://bugzilla.zimbra.com/show_bug.cgi?id=29895

Copied from Perforce
 Change: 107022