(split)bug: 29993
We were using search query to look for expired messages.  There was a logic error that caused the query to run unnecessarily in some cases.  In addition, when a user changes filter, say from 1 month to all, or the other way around, the query route can be very expensive.  This change switches to db query instead, and added a bunch of optimizations.

http://bugzilla.zimbra.com/show_bug.cgi?id=29993

Copied from Perforce
 Change: 109124