(split)bug: 30929
When there's no <refresh> block (either because of <nosession> or because of lack of cross-server notification), ZMailbox now fetches the tag and folder data separately when it's accessed.

GetTag now returns a blank list rather than service.PERM_DENIED when the requester does not have access to the target mailbox's tags.

http://bugzilla.zimbra.com/show_bug.cgi?id=30929

Copied from Perforce
 Change: 112316