(split)bug: 30049
For key grantee type, the accesskey is not encoded in the <notify> block in FolderAction or the <refresh> block for any calls.   accesskey is only returned on explicit GetFolderRequest.   Before this fix, after ZMailbox.modifyFolderGrant is called, cleint still have to call an explicit GetFolderRequest in order to get the accesskey.

add a convenient shortcut that:
     - on the server, returns accesskey in FolderActionResponse body
     - in zclient, updates the ZGrant object from  FolderActionResponse, because the accesskey won't be available in the <notify> in the resposne context so the normal refreshing mechanism won't work.

http://bugzilla.zimbra.com/show_bug.cgi?id=30049

Copied from Perforce
 Change: 112700