(split)bug: 31058
When Outlook/ZCO sends SendMsgRequest with needCalendarSentByFixup=1 and the iCalendar object is a REPLY, lookup the appointment by UID and use ORGANIZER in it to replace what's in the reply.  Or add ORGANIZER if reply didn't have one.  This allows the recipient of the reply to update the appointment in the correct organizer mailbox.

http://bugzilla.zimbra.com/show_bug.cgi?id=31058

Copied from Perforce
 Change: 113395