(split)bug: 30644
Adjust item's current version to 1 greater than the max of the existing revisions before adding a new revision, and log to mailbox.log.  Without this check we may get duplicate row error while inserting to the revision table.  This indicates the version incrementing isn't happening correctly all the time, but I'm not sure where.  I have seen the error happen, so the issue is out there.  The bandaid is good for now since calendar revisioning is only used as a debugging tool.

http://bugzilla.zimbra.com/show_bug.cgi?id=30644

Copied from Perforce
 Change: 113658