(split)bug: none
Backup/restore unit testing.

* Updated ZimbraBackup project to depend on JUnit.
* Added public static versions of methods in BackupUtil, to allow them
to be executed by unit test code.
* Added new TestCreateMessage unit test.  This test delivers a bunch
of messages with sizes that are below and above the compression threshold
and disk streaming threshold.  The test backs up the account, restores
to a new account, and confirms that the blobs on disk are stored in
the correct format.
* Fixed a bug in ByteUtil.isGzipped(), which didn't correctly return
true for gzipped data.
* Fixed JUnit source path in ZimbraServer.
* Updated TestUtil to allow the caller to specify the message body
for generated test messages.

Copied from Perforce
 Change: 113665