(split)bug: 31323
Whenever a Mailbox is deleted from the server (this includes the case of moving the Mailbox from one server to another via the zmmailboxmove command) the WaitSet will return an error of type MAILBOX_DELETED indicating that the Mailbox has been deleted and is no longer a part of the waitset.

<WaitSetResponse waitSet="id"...>
  <error id="account-id" t="MAILBOX_DELETED"/>
  ...possibly other errors...
  ...possibly normal signals...
</WaitSetResponse>

Caller should consult with the directory to see if the Account has been really deleted, of if the Mailbox has been moved to another host.  The account is removed from the waitset, it is the caller's responsibility to re-add it locally (if the Mailbox is recreated) or on another server.

http://bugzilla.zimbra.com/show_bug.cgi?id=31323

Copied from Perforce
 Change: 113767