(split)bug: 31343
For Derby, if the search SQL query we're trying to run has an OR at the top level, run it as N separate individual SQL queries and then merge the result sets in memory.  This hopefully is enough to work around Derby's problems with the conversation search query.

The fix currently only works for sort-by-date, supporting other sorts is a TODO but tricky because of collator issues.

http://bugzilla.zimbra.com/show_bug.cgi?id=31343

Copied from Perforce
 Change: 114168