(split)bug: 27645
Introduced the notion of "local-only" exceptions.  Local-only exceptions are exceptions that exist only in an attendee's appointment.  Such an exception is created when attendees sets a reminder on the instance, for example.  The exception is local-only because it does not exist in the organizer's appointment or any other attendee's.

When organizer sends an email updating the recurrence series, the local-only exceptions are made to "snap" to the series.  That is, the exceptions are re-initialized using the data from the new series.  Exceptions still remain so the custom reminder values can be preserved, but all other fields snap back to the series.

On the other hand, an exception published by the organizer is not local-only and are not made to snap to the series when series is updated.

When ZCO sends a calendar invite email updating recurrence series, set X-ZIMBRA-DISCARD-EXCEPTIONS:TRUE property.  On the attendee side, treat all exceptions as local-only and make them snap to the series.  This effectively makes all exceptions go away.  This is necessary to make attendee's appointment consistent with organizer's appointment edited with Outlook 2003/2007.

In all cases, after recurrence series is modified by the organizer, everyone on the meeting should have the same data.

If the same instance is edited by attendee and organizer, the exception instance is considered not local-only.

All exceptions in the organizer's mailbox are considered not local-only.

When an appointment is ics-exported, any exception set to local-only will have X-ZIMBRA-LOCAL-ONLY:TRUE property.  Import side will honor this flag.

In the item metadata, local-only exceptions will have "lo" field set to true.  Non-local-only exceptions won't set "lo".  Thus, exceptions in existing installs prior to this fix are considered non-local-only.  This is a reasonable default and no data migration is necessary during upgrade.

http://bugzilla.zimbra.com/show_bug.cgi?id=27645

Copied from Perforce
 Change: 115177