(split)bug: 30049
For the new key grantee, for security reason we don't return the access key in the normal <acl> in <refresh> block.  The access key for a key grant is only returned in GetFolderResponse.   Problems arose when user1 grants admin right to account user2, and then user2 grants right of the mountpoint to a key grantee, because we didn't return <acl> for mountpoint in GetFolderResponse so ther is no way to return the accesskey for the grant.    

It now returns <acl> for mountpoint, there is no specific reason not to.  

http://bugzilla.zimbra.com/show_bug.cgi?id=30049

Copied from Perforce
 Change: 115635