(split)bug: 31590
resp.setContentType has to be called *before* resp.getWriter(), or else jetty thinks this response is already in writing(even if nothing has been written to the writer yet) state and will ignore the charset in setContentType call.  It will then use whatever charset is last set (by setCharacterEncoding() or setContentType()).  If none was set, it will use iso-8859-1 as the charset.

http://bugzilla.zimbra.com/show_bug.cgi?id=31590

Copied from Perforce
 Change: 116971