(split)bug: 31629
Move the trashed appointment check out to the email delivery layer, so it doesn't get triggered during a SetAppointment call on Desktop server.

The purpose of the check is to ignore an appointment in trash folder if an update is sent by organizer.  The update should be treated as a new appointment invite, i.e. add an appointment in calendar folder and don't mix it with any existing invites.

http://bugzilla.zimbra.com/show_bug.cgi?id=31629

Copied from Perforce
 Change: 117088