(split)bug: none
New filter rules SOAP API checkpoint.  Added SieveVisitor class
for iterating a Sieve node tree and an initial implementation
(SieveToSoap) that converts a Sieve tree to the new-style
SOAP format.  This code is not being called yet.

Added a DateParser utility class, which is a thread-safe
wrapper around SimpleDateFormat.

Fixed a bug in the regexp used to match filter rule names.
Now it matches the entire rule name, as opposed to just the
first word.

Copied from Perforce
 Change: 117319