(split)bug: 32207
The apache reverse proxy is re-writing the Host header to be the MBS IP.  It sets the original request hostname in the X-Forwarded-Host header.  This breaks virtual host related fratures.  To work around it,  when we grab hostname from the http request for virtual host purpose, we first check for X-Forwarded-Host and then fallback to Host.

http://bugzilla.zimbra.com/show_bug.cgi?id=32207

Copied from Perforce
 Change: 118391