(split)bug: 27630
Do not special case timezone ID "GMT".  While its meaning is well-known and obvious, some calendar systems are known to abuse it by calling a non-GMT VTIMEZONE with the name "GMT".  According to a customer report Lotus Notes does this.  (not sure which version)

We used to special case "GMT" and always map it to the UTC timezone.  By not doing this any more, we can now use the definition supplied in the VTIMEZONE section of an iCalendar object.

http://bugzilla.zimbra.com/show_bug.cgi?id=27630

Copied from Perforce
 Change: 119882