(split)bug: 17999 - Filters: forwarding to oneself results in bounce back to originator
Now adding X-Zimbra-Forwarded headers to messages that
are redirected by mail filters.  If the mail filtering code finds
that the header is set to the current user's account, it now
detects the mail loop and delivers to Inbox instead.

Updated TestUtil LMTP methods to handle user accounts
with and without domains.  Tweaked ParsedMessage.getHeader()
API's so that the return values are more consistent.  When
the header does not exist, getHeader() now returns null and
getHeaders() returns and empty array.

http://bugzilla.zimbra.com/show_bug.cgi?id=17999

Copied from Perforce
 Change: 120228