(split)bug: 32754 - POP3/IMAP tracing should be per-account, not global
Updated Pop3Import and ImapImport to create sessions on
the fly instead of using a single global session object.  If
zimbraDataSourceEnableTrace is set to TRUE, for the data
source, we now call setDebug(true) on the session object
and JavaMail writes the protocol trace to stdout (zmmailboxd.out).

Decided not to mess with SMTP at this point, because the
hooks are not already there.  We can turn on tracing for
SMTP later if we determine that it's necessary.

http://bugzilla.zimbra.com/show_bug.cgi?id=32754

Copied from Perforce
 Change: 121450