bug: 33197
Message file for Hindi are named ***_hi.properties.  hi is not a supported java locale, hi_IN is.     Our resource scan code used to validate that each scanned locale must be a valid java locale, or else it won't be recognized.    This is why hi was not returned.      After talking to Andy, we now honor all scanned locales, regardless if they are supported by java.    i.e. if {base}_foo_BAR.properties exists in the message directory, we now return foo_BAR as a lcoale in GetAvailableLocalesResponse , even if foo is not a language and BAR is not a country.

http://bugzilla.zimbra.com/show_bug.cgi?id=33197

Copied from Perforce
 Change: 123236