(split)bug: 23923
Server now sends unsubscribed roster notifications (with ask="subscribe") to indicate outstanding-but-not-replied-to buddy requests.

<n type="unsubscribed" to="TOADDR" name="NAME" groups="GROUPS" ask="subscribe"/> 

-- this means that the subscription is NOT ACTIVE (you are NOT currently receiving their presence information, but we have an outstanding request to that user and we are waiting for them to respond.

Note that the UI *must* allow the user to re-send the buddy request: the XMPP subscribe protocol is network-reliable and there are cases where the remote user might have lost or not received our buddy request....the user has to be able to force the UI to re-send the buddy request in those cases.

http://bugzilla.zimbra.com/show_bug.cgi?id=23923

Copied from Perforce
 Change: 124093