bug: 33396
Support VTIMEZONE TZNAME property.  The TZNAME is usually the abbreviated name for standard and daylight times in a time zone.  For example, "PST" and "PDT" for US Pacific time zone.  In SOAP API this data is exposed as stdname and dayname attributes of <tz> element.  Clients can use this info in date string formatting.

http://bugzilla.zimbra.com/show_bug.cgi?id=33396

Copied from Perforce
 Change: 124126