bug: 6491 - Allow automatic / scheduled reloading of feeds
Added support for RSS data sources.  When a URL is set on a folder, we now
implicitly create a data source that updates the folder's RSS feed every
12 hours.  Details:

* Added rssDataSource and liveDataSource object classes.
* Updated zimbraDataSourceHost, zimbraDataSourcePort, and
zimbraDatasourceConnectionType to be required in pop3DataSource and
imapDataSource but not dataSource, since they are not required in
rssDataSource.
* Updated zimbraDataSourceEnabled to call the DataSourceCallback
on modifications, so that we update scheduling when data sources
are enabled/disabled.
* Added default COS value of 12h to zimbraDataSourceRssPollingInterval.
* Fixed typo in the spec for <TestDataSourceRequest>.
* Updated DataSourceCallback to handle the new protocol-specific
polling intervals, and data sources being enabled/disabled.
* Updated DataSource.getPollingInterval() to use the new attrs.
* Added migration code to DataSource that sets the IMAP and POP3
polling intervals on an account or cos when the old polling interval
is set.  This will only run once per account/cos.
* Updated DataSourceManager and DataSourceTask to turn off scheduled
polls when the data source is disabled.
* Added RssImport implementation.
* Updated Mailbox code to automatically create/delete an RssDataSource
based on the folder's URL value.
* Added ZRssDataSource class to the ZMailbox API.
* Added unit tests for <TestDataSourceRequest>, RSS data sources,
and polling interval migration.
* Updated TestUtil.deleteTestData() to clean up data sources.

http://bugzilla.zimbra.com/show_bug.cgi?id=6491

Copied from Perforce
 Change: 125453