(split)bug: 19111
Set sequence to the max of 1) what the modifying client specified (if any), 2) current sequence + 1, and 3) current sequence of the series, if we're modifying an exception instance.  This is to satisfy Outlook's requirement that an up-to-date exception instance should have a sequence greater than or equal to the series sequence.

http://bugzilla.zimbra.com/show_bug.cgi?id=19111

Copied from Perforce
 Change: 127321