(split)bug: 27389
added server/globalconfig attr zimbraMailClearTextPasswordEnabled - Whether to allow password sent to non-secured port when zimbraMailMode is mixed.
If it set to TRUE the server will allow login with clear text AuthRequests and change password with clear text ChangePasswordRequest.
If it set to FALSE the server will return an error if an attempt is made to ChangePasswordRequest or AuthRequest.  

http://bugzilla.zimbra.com/show_bug.cgi?id=27389

Copied from Perforce
 Change: 139895