(split)bug: 32254
When a new attendee is added to a recurring appointment's series, add the attendee to all future exceptions.  Multiple emails are sent to the attendee:

- one email containing the series data plus all canceled instances in the future.  The text/calendar part includes one VEVENT  for the series and zero or more cancel VEVENTs.
- one email per non-cancel (i.e. modified) exception instance in the future.

Exceptions in the past are not sent to the new attendee because they are no longer relevant.

Modified instances are not combined into a single email because each may contain different body text and/or attachments.

http://bugzilla.zimbra.com/show_bug.cgi?id=32254

Copied from Perforce
 Change: 141371