(split)bug: 34532
CreateMailbox redo op starts by checking if the mailbox in question already exists.  This check used to rely on the behavior of provisioning layer that threw NO_SUCH_MBOX exception if the mailbox doesn't exist for the account id passed in.  But provisioning behavior changed to either returning null or auto-creating the mailbox, depending on method signature.  With this mismatch, CreateMailbox redo op was doing get-with-auto-create, and the auto create path didn't use the mailbox id saved in the redo op.  Fixed by changing the getMailbox call to the method that doesn't auto create mailbox, and checking for null return value instead of exception.

Also, DbMailbox.createMailbox() call had a bug where the group_id column in mailbox table was being set to 0 when the call is made with an explicit id rather than ID_AUTO_INCREMENT.  This problem is fixed as well.  This bug was manifesting as SQLException complaining that "mboxgroup0" database doesn't exist.  (This was reported in some support cases; mystery solved!)

http://bugzilla.zimbra.com/show_bug.cgi?id=34532

Copied from Perforce
 Change: 141483