bug: 22577
Much more efficient way to update the cached summary when appointments change.  Notify the summary cache whenever an appointment is created/deleted/moved/changed.  Cache will remember the list of items that changed and if there are only a few changed items, next read request will update the cached data with minimal work.  Previously any change in the calendar folder would invalidate the entire cache for that folder and result in refetching/recomputing all appointments.  The fetch from mboxgroupN.appointment table and associated java object creation was quite slow, especially when there are a lot of appointments.

On a busy calendar, this change should result in a big improvement.

If too many (configurable, default 10) appointments changed, update the cache the old way because that may be faster than many separate database queries to fetch modified appointments.

During ics import the cache gets a bunch of update notifications and the list can get too big if import creates many new appointments.  This could be a memory hazard, so just toss the calendar's cache if more than 100 changes accumulate in that folder.

http://bugzilla.zimbra.com/show_bug.cgi?id=22577

Copied from Perforce
 Change: 142763