(split)bug 24889 - Webclient reports inaccurate error for MTA 550 messages
Added the entire error string from the SafeSendFailedException
to the MailServiceException.

We can't know for sure that the sender
is being rejected, because the MTA returns 550 as a response to
the RCPT TO command.  At least now we'll have the entire error
string, so it will be easier to diagnose the problem.

http://bugzilla.zimbra.com/show_bug.cgi?id=24889

Copied from Perforce
 Change: 142784