(split)bug: 33892
There's a new attribute "noNextAlarm" on <SetAppointmentRequest>.  If noNextAlarm="1", this tells the server the last reminder on this appointment has been dismissed and there is no more alarm to trigger.

There are two other cases for SetAppointmentRequest's next alarm handling, and these work as before.  If "nextAlarm" attribute is set, this tells the server to take that value and save it.  (Server actually does some computation to choose an alarm that is at or later than the provided value, but this is off topic for this bug.)  When specifying nextAlarm, the noNextAlarm attribute should not be set.

The final case is omitting both nextAlarm and noNextAlarm.  This tells the server to preserve the current value of next alarm.

The GetAppointmentResponse format remains the same.  If there's a next alarm to trigger, <appt> will have nextAlarm.  If there is none to trigger, nextAlarm will not be set.

During ZDesktop sync, absence of nextAlarm in GetAppointmentResponse is translated into noNextAlarm=1 on SetAppointmentRequest, for both sync directions.

http://bugzilla.zimbra.com/show_bug.cgi?id=33892

Copied from Perforce
 Change: 142890