(split)bug: 22577
When making GetMiniCalRequest and SearchRequest for shared calendars, group the remote accounts by server and send only one proxy request per target server.  Currently each remote account is proxied separately.  When calendar cache is enabled, this HTTP proxying overhead dominates the run time, so minimize it.

For GetMiniCalRequest the grouping is done all the time.  For SearchRequest, it is done only when calendar summary cache is enabled.

http://bugzilla.zimbra.com/show_bug.cgi?id=22577

Copied from Perforce
 Change: 143115