(split)bug: 22577
When shared calendars in SearchRequest or GetMiniCalRequest runs into PERM_DENIED error, don't fail the entire multi-calendar request.  Log the permission error and skip it so that the calendars that have permission can be queried successfully.

When calendar summary cache data loaded from file is outdated, don't just toss it.  It's likely most of the appointments in the folder is up to date, so reuse the data and refresh only the appointment that were actually changed.

http://bugzilla.zimbra.com/show_bug.cgi?id=22577

Copied from Perforce
 Change: 143194