(split)bug: 22577
Fix a problem with calculating the cache time range end time when the last month being cached is a February.  Because the month view and minical displays 6 weeks all the time the cache range must be aligned to this boundary, and depending on the day of week for February 1st this means including up to 2 weeks of March in February month view.  But the old code was only caching up to the first week of march.  This would end up bypassing the cache which is inefficient.

http://bugzilla.zimbra.com/show_bug.cgi?id=22577

Copied from Perforce
 Change: 143220