(split)bug: 29366
Send permission-denied auto-decline email to organizer if both of the following conditions are true:

    1) organizer is not allowed to invite the user
    2) organizer and the invited user are in the same domain
    3) the receiving user was listed as an attendee, i.e. invite wasn't received indirectly through a mailing list or forward

The third condition is there to reduce noise.  This condition can be removed (send the auto reply anyway) by setting a hidden localconfig key debug_calendar_enable_invite_denied_reply_for_unlisted_attendee to true.

The reply email will have:

    Subject: Decline: [meeting title]
    Body: You do not have permission to invite this user to a meeting.

Both are configurable in /opt/zimbra/conf/msgs/ZsMsg.properties, with localization support.  The body text is set with a new message key called calendarUserReplyPermissionDenied.

The reply includes an iCalendar REPLY object with PARTSTAT=DECLINED.  This allows the organizer's calendar to reflect the denial.

Also changed the behavior of the usual auto-accept/decline messages sent by calendar resources.  These emails are now saved to the resource's Sent folder.  To avoid nesting maibox transactions, the email sends are done in a separate thread.

On the organizer side, if a reply is received from a user who isn't a named attendee on the appointment, this new attendee gets added.  This is an existing behavior.  But now we skip the add if the reply was a decline reply.  This makes sense.  If a meeting invites a mailing list and people start sending relies as themselves, it's useful to add the accepting attendees so their status can be seen, but it's not useful to show the declining users and it only generates clutter.  A decline reply from an already-listed attendee is preserved, of course.

http://bugzilla.zimbra.com/show_bug.cgi?id=29366

Copied from Perforce
 Change: 144017