bug: 35259, 32675
Ability for a calendar resource partially accept a recurring appointment when number of conflicts is within configured limit.

Two preferences control the limits:

* zimbraCalResMaxNumConflictsAllowed
* zimbraCalResMaxPercentConflictsAllowed

The first one sets the maximum *number* of conflicts allowed.  The second one sets the maximum *percent* of conflicts allowed.  The smaller of the max num and the number determined by the percent parameter is used to determine whether the reject the meeting request or not.

Only conflicts in the future are counted.  For percent-based calculation, the math is:

# conflicts allowed = (total instances in the future) * (pct configured) / 100

If there are more future conflicts than this, we're over the limit.

Both parameters default to 0, meaning even a single conflict results in rejecting the whole recurring appointment.  This is consistent with the old behavior.

When the number of conflicts is small enough, the resource will do a partial accept.  It updates its own status with ACCEPTED for the recurrence series and DECLINED for each conflicting instance.  It then sends one Accept reply to organizer, with a single iCalendar REPLY part with ACCEPTED participation status.  The email body explains the meeting was accepted but some dates were excluded, with list of those dates.

The resource also sends a second reply email to indicate declines.  The subject shows "Decline: ...", so the organizer gets extra feedback that this appointment wasn't booked completely.  The replay contains iCalendar REPLY object containing multiple VEVENT parts, one for each declined instance.  The email body again shows the declined dates, and also lists the conflicting meetings that are already on resource's calendar.

http://bugzilla.zimbra.com/show_bug.cgi?id=35259
http://bugzilla.zimbra.com/show_bug.cgi?id=32675

Copied from Perforce
 Change: 145311