(split)bug: 34786
When attendee uses ajax client to accept/decline a non-exception instance of a recurring appointment, first create a local exception instance and record the participation status there.  With the full-blown exception instance, it is now possible to see the correct data in ics export and mobile sync clients.

The exception is created as a local-only one, similar to what happens when attendee sets per-instance reminder.

When attendee receives update to the recurrence series, reset the participation status of local-only exceptions to NEEDS-ACTION.  The local-only exceptions are supposed to act like they don't really exist, so it makes sense to reset the partstat to NEEDS-ACTION, the same value the non-exception instances revert to after receiving an update from organizer.

http://bugzilla.zimbra.com/show_bug.cgi?id=34786

Copied from Perforce
 Change: 145384